% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_background_variable.R
\name{add_background_variable}
\alias{add_background_variable}
\title{Add variable names in as background objects}
\usage{
add_background_variable(name)
}
\arguments{
\item{name}{name to store}
}
\description{
Add variable names in as background objects
}
