% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_combined_dataset.R
\name{add_combined_dataset}
\alias{add_combined_dataset}
\title{Combines variables in one dataset}
\usage{
add_combined_dataset(datasets = "all", name)
}
\arguments{
\item{datasets}{names of datasets}

\item{name}{name of combined dataset}
}
\description{
Combines variables in one dataset
}
