% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Adds new data}
\usage{
add_data(
  data,
  name,
  dataset,
  set.default.data.name = F,
  new.observations.set.name
)
}
\arguments{
\item{data}{new data}

\item{name}{data name}

\item{dataset}{dataset}

\item{set.default.data.name}{set data name as default}

\item{new.observations.set.name}{if new observations set is added, specify
name}
}
\description{
Adds new data
}
