% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_database.R
\name{add_database}
\alias{add_database}
\title{Adds new database to .databases}
\usage{
add_database(database, id, type, replace = F)
}
\arguments{
\item{database}{new database}

\item{id}{database id}

\item{type}{database type}

\item{replace}{Should existing database be replaced?}
}
\description{
Adds new database to .databases
}
