% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_columns.R
\name{add_group_columns}
\alias{add_group_columns}
\title{Helper function for df2groupedList to add column/s to the list}
\usage{
add_group_columns(grouped.list, x, from, to, separate)
}
\arguments{
\item{grouped.list}{list to be appended}

\item{x}{source data frame}

\item{from}{start column index}

\item{to}{end column index}

\item{separate}{should columns be added separately}
}
\description{
Helper function for df2groupedList to add column/s to the list
}
