% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_import.R
\name{add_import}
\alias{add_import}
\title{Saves imported files to .imports list}
\usage{
add_import(import)
}
\arguments{
\item{import}{list of imported files}
}
\description{
Saves imported files to .imports list
}
