% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observations_data.R
\name{add_observations_data}
\alias{add_observations_data}
\title{Adds data to observations data frame}
\usage{
add_observations_data(
  data,
  name,
  observations.set,
  dataset,
  order.by = F,
  ignore.names = F,
  add.background.variable = T,
  replace
)
}
\arguments{
\item{data}{new data}

\item{name}{name}

\item{observations.set}{set of observations}

\item{dataset}{dataset}

\item{order.by}{orders observations data by given vector/factor levels}

\item{ignore.names}{Assumes that data matches observations}

\item{add.background.variable}{add name of column to the background
variables for easy access}

\item{replace}{replace existing column}
}
\description{
Adds data to observations data frame
}
