% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observations_set.R
\name{add_observations_set}
\alias{add_observations_set}
\title{Adds a new set of observations}
\usage{
add_observations_set(name, observations, dataset, set.default = T)
}
\arguments{
\item{observations}{vector of observations}

\item{dataset}{dataset}

\item{set.default}{set new as default}

\item{set}{name for new set of observations}
}
\description{
Adds a new set of observations
}
