% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_datasets.R
\name{add_raw_dataset}
\alias{add_raw_dataset}
\title{Saves raw data frames to .info list}
\usage{
add_raw_dataset(raw_dataset, name)
}
\arguments{
\item{raw_dataset}{raw data frame}

\item{name}{name of dataset}
}
\description{
Saves raw data frames to .info list
}
