% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_segment_pvalue.R
\name{add_segment_pvalue}
\alias{add_segment_pvalue}
\title{Adds p-values to geom_segment dataframe fo rdendrogram plot}
\usage{
add_segment_pvalue(segment, dend.enrich)
}
\arguments{
\item{segment}{segment data}

\item{dend.enrich}{dend.table.enrichment}
}
\description{
Adds p-values to geom_segment dataframe fo rdendrogram plot
}
