% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables_data.R
\name{add_variables_data}
\alias{add_variables_data}
\title{Add data to variables}
\usage{
add_variables_data(
  data,
  name,
  dataset,
  set.default = F,
  add.background.variable = T,
  replace
)
}
\arguments{
\item{data}{new data}

\item{name}{name}

\item{dataset}{dataset}

\item{set.default}{set new variables data as default}

\item{add.background.variable}{add name of column to the background
variables for easy access}

\item{replace}{replace existing column}
}
\description{
Add data to variables
}
