% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Calculates adjacency matrix from similarity matrix}
\usage{
adjacency(data, method, alpha, theta, beta)
}
\arguments{
\item{data}{similarity matrix}

\item{method}{adjacency function ("sigmoid", "power", "none")}

\item{alpha}{alpha parameter for sigmoid function}

\item{theta}{theta parameter for sigmoid function}

\item{beta}{beta parameter for power function}
}
\description{
Calculates adjacency matrix from similarity matrix
}
