% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_layer_discrete_x.R
\name{annot_layer_discrete_x}
\alias{annot_layer_discrete_x}
\title{Creates discrete annotation layer}
\usage{
annot_layer_discrete_x(
  data,
  pos_table_x,
  x = "observations",
  value = "groups",
  colors = NULL,
  legend.direction = "vertical",
  legend.title = ""
)
}
\arguments{
\item{data}{data}

\item{pos_table_x}{pos_table_x (function specific)}

\item{x}{x-axis column}

\item{value}{value column}

\item{colors}{color vector}

\item{legend.direction}{direction of legend entries ("horizontal", "vertical")}

\item{legend.title}{title of legend}
}
\description{
Creates discrete annotation layer
}
