% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_data2list.R
\name{annotation_data2list}
\alias{annotation_data2list}
\title{Transforms data frame to annotation list}
\usage{
annotation_data2list(data.annotation, sep = "; ")
}
\arguments{
\item{data.annotation}{annotation data from UniProt.ws}

\item{sep}{separator}
}
\description{
Transforms data frame to annotation list
}
