% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_name.R
\name{ask_name}
\alias{ask_name}
\title{Asks user for name if no name given}
\usage{
ask_name(name, message, exclude = NA)
}
\arguments{
\item{name}{name}

\item{message}{message}

\item{exclude}{exclusion list for unavailable names}
}
\description{
Asks user for name if no name given
}
