% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_heatmap_rowwise.R
\name{assemble_heatmap_rowwise}
\alias{assemble_heatmap_rowwise}
\title{Combines subplots to form a heatmap}
\usage{
assemble_heatmap_rowwise(
  plot.list,
  rel_dend_x,
  rel_dend_y,
  rel_labels_x,
  rel_labels_y,
  rel_annot_layer_x,
  rel_annot_layer_y,
  rel_legends,
  rel_legends_space
)
}
\arguments{
\item{plot.list}{list of plots from plot_gg_heatmap()}

\item{rel_dend_x}{relative height of x-axis dendrogram}

\item{rel_dend_y}{relative width of y-axis dendrogram}

\item{rel_labels_x}{relative height of x-axis labels}

\item{rel_labels_y}{relative width of y-axis labels}

\item{rel_annot_layer_x}{relative height/s of x-axis annotation layers (either one number or vector with number for each)}

\item{rel_annot_layer_y}{relative width/s of y-axis annotation layers (either one number or vector with number for each)}

\item{rel_legends}{relative width of legends area}

\item{rel_legends_space}{realtive space between plot and legends}
}
\description{
Combines subplots to form a heatmap
}
