% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data_frames.R
\name{available_data_frames}
\alias{available_data_frames}
\title{Shows available raw data frames and returns them or prints a transfer call}
\usage{
available_data_frames(dataset, view = T, return = F, print.call = F)
}
\arguments{
\item{dataset}{dataset}

\item{view}{View?}

\item{return}{return data frame names}

\item{print.call}{print call for transferring data frames}
}
\description{
Shows available raw data frames and returns them or prints a transfer call
}
