% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_variables_data.R
\name{available_variables_data}
\alias{available_variables_data}
\title{Shows available raw variables data and returns them or prints a transfer call}
\usage{
available_variables_data(dataset, view = T, return = F, print.call = F)
}
\arguments{
\item{dataset}{dataset}

\item{view}{View?}

\item{return}{return all variables data names}

\item{print.call}{print call for transferring variables data}
}
\description{
Shows available raw variables data and returns them or prints a transfer call
}
