% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_limit_breaks.R
\name{axis_limit_breaks}
\alias{axis_limit_breaks}
\title{Helps make nice axis limit breaks}
\usage{
axis_limit_breaks(plot_limits, break.space = 1)
}
\arguments{
\item{plot_limits}{vector containing plot limits}

\item{break.space}{space between breaks}
}
\description{
Helps make nice axis limit breaks
}
