% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhc.R
\name{bhc}
\alias{bhc}
\title{Performs Benjamini-Hochberg-Correction of set of p-values}
\usage{
bhc(p.values, FDR = 0.05)
}
\arguments{
\item{p.values}{named vector of pvalues}

\item{FDR}{false discovery rate}
}
\description{
Performs Benjamini-Hochberg-Correction of set of p-values
}
