% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_msigdbr.R
\name{browse_msigdbr}
\alias{browse_msigdbr}
\title{Opens MSigDB entries}
\usage{
browse_msigdbr(names = "MODULE_212")
}
\arguments{
\item{names}{vector of entry names}
}
\description{
Opens MSigDB entries
}
