% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_proteins.R
\name{browse_proteins}
\alias{browse_proteins}
\title{Opens UniProt page of given protein}
\usage{
browse_proteins(proteins = "Q63HQ2", limit = 10)
}
\arguments{
\item{proteins}{vector of protein accession Ids}

\item{limit}{limit of proteins to be searched}
}
\description{
Opens UniProt page of given protein
}
