% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cor_list.R
\name{build_cor_list}
\alias{build_cor_list}
\title{Builds cor_list object with default parameters}
\usage{
build_cor_list(
  data,
  min.0 = 0.5,
  normalize.method = "pqn",
  shift = 1.8,
  width = 0.3,
  seed = 123,
  similarity.method = "preserve",
  adjacency.method = "none",
  add.connectivity = F,
  connectivity.k = 20,
  connectivity.scale = T,
  plot = F
)
}
\arguments{
\item{data}{data}

\item{min.0}{threshold to keep proteins}

\item{normalize.method}{method for normalization}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed}{seed for imputation}

\item{similarity.method}{method to calculate similarity matrix}

\item{adjacency.method}{method to calculate adjacency matrix}

\item{add.connectivity}{calculate connectivity measures}

\item{connectivity.k}{max number of clusters for connectivity}

\item{connectivity.scale}{scale connectivity values to cluster size}

\item{plot}{print plot of correlation matrix}
}
\description{
Builds cor_list object with default parameters
}
