% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_variables_manually.R
\name{change_variables_manually}
\alias{change_variables_manually}
\title{Allows to manually change the variables data}
\usage{
change_variables_manually(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Allows to manually change the variables data
}
