% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_database.R
\name{check_database}
\alias{check_database}
\title{Checks if database is available}
\usage{
check_database(id, type)
}
\arguments{
\item{id}{database name/identifier}

\item{type}{type of database}
}
\description{
Checks if database is available
}
