% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_terms.R
\name{choose_terms}
\alias{choose_terms}
\title{Interface for user to define a list of GO terms}
\usage{
choose_terms(proteins, TERM2GENE, graphics)
}
\arguments{
\item{TERM2GENE}{TERM2GENE dataframe}

\item{graphics}{Choose from graphic interface}

\item{variables}{variables}
}
\description{
Interface for user to define a list of GO terms
}
