% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_inner_borders.R
\name{cluster_inner_borders}
\alias{cluster_inner_borders}
\title{Returns borders between clusters}
\usage{
cluster_inner_borders(dend, k)
}
\arguments{
\item{dend}{dendrogram (hclust object)}

\item{k}{number of clusters}
}
\description{
Returns borders between clusters
}
