% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_class.R
\name{colnames_class}
\alias{colnames_class}
\title{Returns column names from data frame of specific column class}
\usage{
colnames_class(data, data.class = "numeric")
}
\arguments{
\item{data}{data frame}

\item{data.class}{data type/s}
}
\description{
Returns column names from data frame of specific column class
}
