% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_typeof.R
\name{colnames_typeof}
\alias{colnames_typeof}
\title{Returns column names from data frame of specific column type}
\usage{
colnames_typeof(data, data.type = "double")
}
\arguments{
\item{data}{data frame}

\item{data.type}{data type/s}
}
\description{
Returns column names from data frame of specific column type
}
