% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_prefix.R
\name{common_prefix}
\alias{common_prefix}
\title{Finds common substring; returns "" if no characters are in comomn}
\usage{
common_prefix(names)
}
\arguments{
\item{names}{vector of strings}
}
\description{
Finds common substring; returns "" if no characters are in comomn
}
