% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_enrichment_data.R
\name{compare_enrichment_data}
\alias{compare_enrichment_data}
\title{Compares enrichment data and returns list of common and unique terms}
\usage{
compare_enrichment_data(enrichment.data.x, enrichment.data.y, ...)
}
\arguments{
\item{enrichment.data.x}{enrichment data frame 1}

\item{enrichment.data.y}{enrichment data frame 2}

\item{...}{arguments for simplify_enrichment_results}
}
\description{
Compares enrichment data and returns list of common and unique terms
}
