% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{connectivity}
\alias{connectivity}
\title{Variable-wise connectivity}
\usage{
connectivity(adjacency.matrix, scale = T)
}
\arguments{
\item{adjacency.matrix}{adjacency matrix}

\item{scale}{should connectivity be divided by n - 1}
}
\description{
Variable-wise connectivity
}
