% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy2global.R
\name{copy2global}
\alias{copy2global}
\title{Assigns all arguments with names to Global Environment}
\usage{
copy2global(...)
}
\arguments{
\item{...}{function arguments}
}
\description{
Assigns all arguments with names to Global Environment
}
