% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_log2fc_anova.R
\name{correct_log2fc_anova}
\alias{correct_log2fc_anova}
\title{Calculates correct log2 fold-changes to replace anova diff values}
\usage{
correct_log2fc_anova(posthoc.diff, data)
}
\arguments{
\item{posthoc.diff}{extracted output from TukeyHSD}

\item{data}{data used for anova}
}
\description{
Calculates correct log2 fold-changes to replace anova diff values
}
