% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_variables_per_observation.R
\name{count_variables_per_observation}
\alias{count_variables_per_observation}
\title{Performs analysis}
\usage{
count_variables_per_observation(data_, data.name = "raw_data", plot = T)
}
\arguments{
\item{data_}{data list}

\item{data.name}{name of data to use}

\item{plot}{results}
}
\description{
Performs analysis
}
