% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree_.R
\name{cutree_}
\alias{cutree_}
\title{Like cutree() but returns character vector}
\usage{
cutree_(tree, k = NULL, h = NULL)
}
\arguments{
\item{tree}{dendrogram}

\item{k}{number of cluster}

\item{h}{height of cluster cut}
}
\description{
Like cutree() but returns character vector
}
