% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling_helpers.R
\name{data.frame2tibble}
\alias{data.frame2tibble}
\title{Transforms data frames to tibble and adds column for row names}
\usage{
data.frame2tibble(data, to.row.names)
}
\arguments{
\item{data}{data frame with row names}

\item{to.row.names}{name for row names vector}
}
\description{
Transforms data frames to tibble and adds column for row names
}
