% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2list.R
\name{data2list}
\alias{data2list}
\title{Pull names vector from}
\usage{
data2list(data_, names, input, output, output.type)
}
\arguments{
\item{data_}{list or tibble}

\item{names}{column for vector names; by default first column}

\item{input}{name of input data}

\item{output}{name of output data}

\item{output.type}{(optional) type of output ("list" to force list output)}
}
\description{
Pull names vector from
}
