% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram}
\alias{dendrogram}
\title{Computes a dendrogram from adjacency matrix in cor_object}
\usage{
dendrogram(cor_list, data = "adjacency", method = "average")
}
\arguments{
\item{cor_list}{cor_list object}

\item{data}{data to use for clustering}

\item{method}{clustering method}
}
\description{
Computes a dendrogram from adjacency matrix in cor_object
}
