% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram2table.R
\name{dendrogram2table}
\alias{dendrogram2table}
\title{Adds table representation of dendrogram}
\usage{
dendrogram2table(data_, dendrogram = "dend_x")
}
\arguments{
\item{data_}{data list}
}
\description{
Adds table representation of dendrogram
}
