% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_quantiles.R
\name{distribute_quantiles}
\alias{distribute_quantiles}
\title{Returns named vector and substitutes numeric values by quantile groups}
\usage{
distribute_quantiles(x, n = 4)
}
\arguments{
\item{x}{named numeric vector}

\item{n}{number of quantiles}
}
\description{
Returns named vector and substitutes numeric values by quantile groups
}
