% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling.R
\name{do_column_summary}
\alias{do_column_summary}
\title{Collapses rows of data frame by specified function and adds as list entry}
\usage{
do_column_summary(data_, FUN = sum, name, modify, ignore, input, output)
}
\arguments{
\item{data_}{data}

\item{FUN}{function to combine groups with (default sum)}

\item{name}{(optional) name of new column}

\item{modify}{(optional) character vector of columns to modify}

\item{ignore}{(optional) character vector of columns to ignore}

\item{input}{name of input data frame}

\item{new.name}{name of output data frame (defaults to input data name)}
}
\description{
Collapses rows of data frame by specified function and adds as list entry
}
