% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling.R
\name{do_expr}
\alias{do_expr}
\title{Evaluates data cell-wise}
\usage{
do_expr(data_, expr, FUN, modify, ignore, input, output)
}
\arguments{
\item{data_}{data list}

\item{expr}{expression to be applied to each cell (must contain x as
variable, e.g. x > 2)}

\item{FUN}{function to be applied to each cell}

\item{modify}{(optional) character vector of columns to modify}

\item{ignore}{(optional) character vector of columns to ignore}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\description{
Evaluates data cell-wise
}
