% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_normalize.R
\name{do_normalize}
\alias{do_normalize}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_normalize(data_, method = "pqn", dataset, input, output)
}
\arguments{
\item{data_}{list or tibble}

\item{method}{normalisation method (see normalize())}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
