% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling.R
\name{do_nothing}
\alias{do_nothing}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_nothing(data_, ..., dataset, input, output)
}
\arguments{
\item{data_}{list or tibble}

\item{...}{specific arguments}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
