% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_pca(data_, scale = T, dataset, input, output = "data_pca")
}
\arguments{
\item{data_}{list or tibble}

\item{scale}{scale data}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
