% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_scale.R
\name{do_scale}
\alias{do_scale}
\title{Scales data (Z-scores) from tibbles or analysis_list containing a tibble}
\usage{
do_scale(data_, input, output)
}
\arguments{
\item{data_}{data_ list}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Scales data (Z-scores) from tibbles or analysis_list containing a tibble
}
