% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_t.test.R
\name{do_t.test}
\alias{do_t.test}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_t.test(
  data_,
  paired = F,
  var.equal = T,
  group.column = "groups",
  control.group,
  pAdjustMethod = "BH",
  p.value.cutoff = 0.05,
  fc.threshold = 0,
  input,
  output = "data_t.test"
)
}
\arguments{
\item{data_}{list or tibble}

\item{paired}{a logical indicating whether you want a paired t-test (see ?t.test)}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal (see ?t.test)}

\item{group.column}{column to use as group identification}

\item{pAdjustMethod}{see p.adjust.methods}

\item{input}{name of input}

\item{output}{name of output}
}
\description{
Template for functions that accept either a data frame or a list
}
