% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling.R
\name{do_transpose}
\alias{do_transpose}
\title{Transposes tibble and uses first column as column names}
\usage{
do_transpose(data_, from.row.names, to.row.names, input, output)
}
\arguments{
\item{data_}{data_ list}

\item{from.row.names}{row names column of initial data frame}

\item{to.row.names}{row names column after transposing}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Transposes tibble and uses first column as column names
}
