% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_code_helpers.R
\name{.template_names}
\alias{.template_names}
\title{Lists available template names from pOmics-templates repository}
\usage{
.template_names()
}
\description{
Lists available template names from pOmics-templates repository
}
