% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_attributes.R
\name{.transpose_data_attributes}
\alias{.transpose_data_attributes}
\title{Get column names containing data independently if variables or data columns}
\usage{
.transpose_data_attributes(data_attributes)
}
\arguments{
\item{data_attributes}{data_attributes list}
}
\description{
Get column names containing data independently if variables or data columns
}
