% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_attributes.R
\name{.update_data_attributes}
\alias{.update_data_attributes}
\title{Compares data attributes with data and removes non-existing entries}
\usage{
.update_data_attributes(data)
}
\arguments{
\item{data}{data object with specific attributes}
}
\description{
Compares data attributes with data and removes non-existing entries
}
