% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_code_helpers.R
\name{download_template}
\alias{download_template}
\title{Downloads template scripts and writes to new file}
\usage{
download_template(folder = "Scripts")
}
\arguments{
\item{folder}{where to save template script}
}
\description{
Downloads template scripts and writes to new file
}
