% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_DT_fun_enrich}
\alias{export_DT_fun_enrich}
\title{Formats table for nice output (R Markdown)}
\usage{
export_DT_fun_enrich(data, buttons, dom = "lBfrtip")
}
\arguments{
\item{data}{data from fun_enrich (enrichResult)}

\item{buttons}{(optional) add buttons ("copy", "csv", "excel", "pdf",
"print")}

\item{dom}{order of table elements (default: "lBfrtip", see
https://rstudio.github.io/DT/)}
}
\description{
Formats table for nice output (R Markdown)
}
