% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_GO_table}
\alias{export_GO_table}
\title{Prepares GO table for output in HTML file}
\usage{
export_GO_table(table, caption = NULL)
}
\arguments{
\item{table}{GO results table}

\item{caption}{table caption}
}
\description{
Prepares GO table for output in HTML file
}
