% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_info_template.R
\name{export_info_template}
\alias{export_info_template}
\title{Exports observations names to csv info template file}
\usage{
export_info_template(dataset, path, observations.set)
}
\arguments{
\item{dataset}{dataset}

\item{path}{directory}

\item{observations.set}{observations.set}
}
\description{
Exports observations names to csv info template file
}
