% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_RMD_fun_enrich}
\alias{export_RMD_fun_enrich}
\title{Template for functions that accept either a data frame or a list}
\usage{
export_RMD_fun_enrich(
  data_,
  buttons,
  dom = "lBfrtip",
  view = F,
  input = "fun_enrich",
  output = "fun_enrich_DT"
)
}
\arguments{
\item{data_}{list or tibble}

\item{buttons}{(optional) add buttons ("copy", "csv", "excel", "pdf",
"print")}

\item{dom}{order of table elements (default: "lBfrtip", see
https://rstudio.github.io/DT/)}

\item{view}{view table}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
