% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pdf.R
\name{export_svg}
\alias{export_svg}
\title{Exports plot to svg}
\usage{
export_svg(p, file = "test.svg", width = 4, height = 4, open = T)
}
\arguments{
\item{p}{plot}

\item{file}{filename}

\item{width}{width}

\item{height}{height}

\item{open}{open pdf after export}
}
\description{
Exports plot to svg
}
