% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mqpar_info.R
\name{extract_mqpar_info}
\alias{extract_mqpar_info}
\title{Extracts information from mqpar.xml file}
\usage{
extract_mqpar_info(file)
}
\arguments{
\item{file}{path to mqpar.xml file}
}
\description{
Extracts information from mqpar.xml file
}
