% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_matrix.R
\name{factorize_matrix}
\alias{factorize_matrix}
\title{Divides each row of a matrix by the elements of the specified row}
\usage{
factorize_matrix(data, row)
}
\arguments{
\item{data}{expression data}

\item{row}{reference row}
}
\description{
Divides each row of a matrix by the elements of the specified row
}
