% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data_entry.R
\name{find_data_entry}
\alias{find_data_entry}
\title{Identifies data entrys by column names}
\usage{
find_data_entry(data, entry, data.origin)
}
\arguments{
\item{data}{rawdata list}

\item{entry}{entry type}

\item{data.origin}{(optional) data origin}
}
\description{
Identifies data entrys by column names
}
