% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findORchoose.R
\name{findORchoose}
\alias{findORchoose}
\title{Finds vector entry based on given patterns and lets User choose if not possible}
\usage{
findORchoose(names, patterns, title = NULL)
}
\arguments{
\item{names}{vector of strings}

\item{patterns}{vector containing patterns}

\item{title}{Title of user interface}
}
\description{
Finds vector entry based on given patterns and lets User choose if not possible
}
