% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Computes the geometric mean}
\usage{
geomean(x, na.rm = T)
}
\arguments{
\item{x}{vector}

\item{na.rm}{should NAs be removed?}
}
\description{
Computes the geometric mean
}
