% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_terms.R
\name{get_child_terms}
\alias{get_child_terms}
\title{Returns all offspring terms of given GO terms}
\usage{
get_child_terms(terms)
}
\arguments{
\item{terms}{GO terms}
}
\description{
Returns all offspring terms of given GO terms
}
