% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_cut.R
\name{get_cluster_cut}
\alias{get_cluster_cut}
\title{Returns group information of specified number of cluster}
\usage{
get_cluster_cut(cor_list, k = 1)
}
\arguments{
\item{cor_list}{cor_list object}

\item{k}{number of clusters}
}
\description{
Returns group information of specified number of cluster
}
