% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_proteins.R
\name{get_cluster_proteins}
\alias{get_cluster_proteins}
\title{Returns vector of proteins}
\usage{
get_cluster_proteins(cor_list, k = 1, l = 1)
}
\arguments{
\item{cor_list}{cor_list object}

\item{k}{number of clusters}

\item{l}{index of cluster}
}
\description{
Returns vector of proteins
}
