% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_data.R
\name{get_cor_data}
\alias{get_cor_data}
\title{Returns data from cor_list object}
\usage{
get_cor_data(cor_list, name = "adjacency", k = 1, l = 1)
}
\arguments{
\item{cor_list}{cor_list object}

\item{name}{data type (any of "raw_data", "correlation", "similarity", "adjacency")}

\item{k}{cluster cut k}

\item{l}{cluster index l}
}
\description{
Returns data from cor_list object
}
