% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{get_data}
\alias{get_data}
\title{Assemble data from dataset and return in list}
\usage{
get_data(
  which,
  variables = "default",
  observations = "default",
  output,
  output.type = "tibble_inlist",
  observations.set,
  dataset
)
}
\arguments{
\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{output}{name to save in list}

\item{output.type}{output type (default = "tibble_inlist"; combination of
"tibble", "data.frame" or "matrix" to define data type and "_inlist" or ""
if data should be put into a list or not)}

\item{observations.set}{set of observations}

\item{dataset}{dataset name or number}
}
\description{
Assemble data from dataset and return in list
}
