% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_template_names.R
\name{get_data_template_names}
\alias{get_data_template_names}
\title{Returns existing templates for raw data}
\usage{
get_data_template_names()
}
\description{
Returns existing templates for raw data
}
