% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_attr.R
\name{get_dataset_attr}
\alias{get_dataset_attr}
\title{Returns dataset attribute; does not check if exists}
\usage{
get_dataset_attr(which = "names", dataset)
}
\arguments{
\item{which}{attribute type}

\item{dataset}{dataset}
}
\description{
Returns dataset attribute; does not check if exists
}
