% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defaults.R
\name{get_defaults}
\alias{get_defaults}
\title{Returns default data}
\usage{
get_defaults(data.origin, type, silent = F)
}
\arguments{
\item{data.origin}{data origin program}

\item{type}{info type}

\item{silent}{suppress messages}
}
\description{
Returns default data
}
