% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_proteins.R
\name{get_GO_proteins}
\alias{get_GO_proteins}
\title{Returns GO terms from given terms}
\usage{
get_GO_proteins(terms, include.child.terms = F)
}
\arguments{
\item{terms}{GO ids}

\item{include.child.terms}{include child terms}
}
\description{
Returns GO terms from given terms
}
