% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GOdata.R
\name{get_GOdata}
\alias{get_GOdata}
\title{Searches for GOdata or makes new}
\usage{
get_GOdata(proteins, ontology)
}
\arguments{
\item{proteins}{protein vector}

\item{ontology}{GO ontology}

\item{statistic}{statistic}
}
\description{
Searches for GOdata or makes new
}
