% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_column.R
\name{get_labels_column}
\alias{get_labels_column}
\title{Checks default observations labels of no column given}
\usage{
get_labels_column(data, labels, dataset)
}
\arguments{
\item{data}{data frame}

\item{labels}{column name of labels}

\item{dataset}{dataset}
}
\description{
Checks default observations labels of no column given
}
