% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MaxQuant_defaults.R
\name{get_MaxQuant_defaults}
\alias{get_MaxQuant_defaults}
\title{Returns default list for identifier and data types to import}
\usage{
get_MaxQuant_defaults(data.type)
}
\arguments{
\item{data.type}{datatype}
}
\description{
Returns default list for identifier and data types to import
}
