% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Return observations}
\usage{
get_observations(observations, observations.set, dataset)
}
\arguments{
\item{observations}{vector of observations}

\item{observations.set}{observations data frame}

\item{dataset}{dataset}
}
\description{
Return observations
}
