% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_data_names.R
\name{get_observations_data_names}
\alias{get_observations_data_names}
\title{Returns observations data names}
\usage{
get_observations_data_names(observations.set, dataset)
}
\arguments{
\item{observations.set}{observations set}

\item{dataset}{dataset}
}
\description{
Returns observations data names
}
