% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_sets.R
\name{get_observations_sets}
\alias{get_observations_sets}
\title{Returns observations sets}
\usage{
get_observations_sets(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Returns observations sets
}
