% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_template.R
\name{get_observations_template}
\alias{get_observations_template}
\title{Creates template for new variables data}
\usage{
get_observations_template(observations.set, dataset, fill = NA)
}
\arguments{
\item{observations.set}{set of observations}

\item{dataset}{dataset}

\item{fill}{entries of vector}
}
\description{
Creates template for new variables data
}
