% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OrgDb.R
\name{get_OrgDb}
\alias{get_OrgDb}
\title{Checks organism database name from Bioconductor and if package can be loaded}
\usage{
get_OrgDb(OrgDb, dataset)
}
\arguments{
\item{OrgDb}{name of organism database}

\item{dataset}{dataset}
}
\description{
Checks organism database name from Bioconductor and if package can be loaded
}
