% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_list.R
\name{get_protein_list}
\alias{get_protein_list}
\title{Returns lists of proteins}
\usage{
get_protein_list(name)
}
\arguments{
\item{name}{name of protein list}
}
\description{
Returns lists of proteins
}
