% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteins_TERM2GENE.R
\name{get_proteins_TERM2GENE}
\alias{get_proteins_TERM2GENE}
\title{Retrieves proteins from annotation terms}
\usage{
get_proteins_TERM2GENE(proteins, terms, TERM2GENE, graphics)
}
\arguments{
\item{proteins}{subset of proteins to choose from}

\item{terms}{annotation terms}

\item{TERM2GENE}{TERM2GENE dataframe}

\item{graphics}{choose terms from graphical window or console}
}
\description{
Retrieves proteins from annotation terms
}
