% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome.R
\name{get_proteome}
\alias{get_proteome}
\title{Returns all proteome accession numbers of given sepcies}
\usage{
get_proteome(id, dataset)
}
\arguments{
\item{id}{id of proteome}

\item{dataset}{dataset}
}
\description{
Returns all proteome accession numbers of given sepcies
}
