% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_dataset.R
\name{get_raw_dataset}
\alias{get_raw_dataset}
\title{Returns raw datasets from .info list}
\usage{
get_raw_dataset(name)
}
\arguments{
\item{name}{name of dataset}
}
\description{
Returns raw datasets from .info list
}
