% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxId.R
\name{get_taxId}
\alias{get_taxId}
\title{Identifies taxonomy Id if possible, ask otherwise}
\usage{
get_taxId(x)
}
\arguments{
\item{x}{data frame to be inspected}
}
\description{
Identifies taxonomy Id if possible, ask otherwise
}
