% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(which, variables, output.type = "vector", ..., dataset)
}
\arguments{
\item{which}{which variables data to pull (multiple supported)}

\item{variables}{(optional) vector of variables or expression}

\item{output.type}{output type (default = "vector" or "tibble_inlist" for
multiple arguments; combination of "vector", "list", "tibble", or
"data.frame" to define data type and "_inlist" or "" if data should be put
into a list or not)}

\item{...}{arguments for FUN function}

\item{dataset}{dataset}
}
\description{
Return variables data
}
