% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_observations.R
\name{identify_observations}
\alias{identify_observations}
\title{Guesses observations names from raw_dataset}
\usage{
identify_observations(x, pattern = ".")
}
\arguments{
\item{x}{data frame}

\item{pattern}{separation pattern of column names}
}
\description{
Guesses observations names from raw_dataset
}
