% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_separator.R
\name{identify_separator}
\alias{identify_separator}
\title{Identifies most common character from dataset and tests for numbers or letters}
\usage{
identify_separator(x)
}
\arguments{
\item{x}{dataset}
}
\description{
Identifies most common character from dataset and tests for numbers or letters
}
