% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_taxId.R
\name{identify_taxId}
\alias{identify_taxId}
\title{Checks protein Ids for their species}
\usage{
identify_taxId(x, try = 1, silent = F)
}
\arguments{
\item{x}{vector of identifiers}

\item{try}{counter}

\item{silent}{should message be printed when Id found}
}
\description{
Checks protein Ids for their species
}
