% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_variables.R
\name{identify_variables}
\alias{identify_variables}
\title{Finds feasible identifier columns}
\usage{
identify_variables(x, identifier, modify.identifiers = "split", sep)
}
\arguments{
\item{x}{data frame}

\item{identifier}{vector of identifier columns}

\item{modify.identifiers}{action to perform on identifier column ("not" for
nothing; "split" to separate strings by separator)}

\item{sep}{separator}
}
\description{
Finds feasible identifier columns
}
