% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_datasets.R
\name{import2grouped_data}
\alias{import2grouped_data}
\title{Stores data frame in a list and groups adjacent columns}
\usage{
import2grouped_data(import)
}
\arguments{
\item{import}{imported data frame}
}
\description{
Stores data frame in a list and groups adjacent columns
}
