% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_datasets.R
\name{import2raw_dataset}
\alias{import2raw_dataset}
\title{Extracts variables and grouped data from imported data frame}
\usage{
import2raw_dataset(import, identifier, modify.identifiers = "split", add = T)
}
\arguments{
\item{import}{imported data frame}

\item{identifier}{columns to use as identifiers}

\item{modify.identifiers}{action to perform on identifier column ("not" for
nothing; "split" to separate strings by separator)}

\item{add}{add raw_dataset to .info list}
}
\description{
Extracts variables and grouped data from imported data frame
}
