% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_project_setup.R
\name{import_fasta}
\alias{import_fasta}
\title{Imports fasta file and saves proteome}
\usage{
import_fasta(
  file,
  name,
  save.proteome = T,
  save.info = F,
  save.sequences = F,
  replace = F
)
}
\arguments{
\item{file}{fasta file path}

\item{name}{(optional) name of data base entry}

\item{save.proteome}{add all protein accessions as reference proteome}

\item{save.info}{extract all information from fasta headers}

\item{save.sequences}{extract protein sequences}

\item{replace}{replace existing databases}
}
\description{
Imports fasta file and saves proteome
}
