% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Top100_EV.R
\name{import_Top100_EV}
\alias{import_Top100_EV}
\title{Downloads Top100 protein list and converts to protein accession ids}
\usage{
import_Top100_EV()
}
\description{
Downloads Top100 protein list and converts to protein accession ids
}
