% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missForest.R
\name{impute_missForest}
\alias{impute_missForest}
\title{Imputes missing values based on a Random Forest algorithm}
\usage{
impute_missForest(data_, input, output, ...)
}
\arguments{
\item{data_}{list or tibble}

\item{input}{name of input data}

\item{output}{name of output data}

\item{...}{arguments for missForest function}
}
\description{
Imputes missing values based on a Random Forest algorithm
}
