% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_processing.R
\name{impute_norm}
\alias{impute_norm}
\title{Imputes values based on normal distribution}
\usage{
impute_norm(data_, shift = 1.8, width = 0.3, seed = 123, input, output)
}
\arguments{
\item{data_}{data (either a tibble, data frame or matrix)}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed}{seed}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\description{
Imputes values based on normal distribution
}
