% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{include_groups}
\alias{include_groups}
\title{Adds group vector to data frame}
\usage{
include_groups(
  data_,
  groups = "default",
  control,
  observations.set,
  dataset,
  input,
  output
)
}
\arguments{
\item{data_}{analysis list}

\item{groups}{groups}

\item{control}{(optional) which group is control}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Adds group vector to data frame
}
