% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{include_observations_data}
\alias{include_observations_data}
\title{Adds group vector to data frame}
\usage{
include_observations_data(
  data_,
  which,
  name,
  observations.set,
  dataset,
  input,
  output
)
}
\arguments{
\item{data_}{data frame}

\item{which}{observations data}

\item{name}{(optional) name of new column}

\item{observations.set}{which observations.set to use}

\item{dataset}{dataset}

\item{input}{data frame to be modified}

\item{output}{data frame to return in list}
}
\description{
Adds group vector to data frame
}
