% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{include_variables_data}
\alias{include_variables_data}
\title{Adds group vector to data frame}
\usage{
include_variables_data(data_, which, name, dataset, input, output)
}
\arguments{
\item{data_}{data frame}

\item{which}{variables data}

\item{name}{name of new column}

\item{dataset}{dataset}

\item{input}{data frame to be modified}

\item{output}{data frame to return in list}
}
\description{
Adds group vector to data frame
}
