% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{initialize_data_structure}
\alias{initialize_data_structure}
\title{Creates basic data structure}
\usage{
initialize_data_structure(replace = F)
}
\arguments{
\item{replace}{replace existing lists}
}
\description{
Creates basic data structure
}
