% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_m.R
\name{intersect_m}
\alias{intersect_m}
\title{Intersect multiple vectors}
\usage{
intersect_m(x, y, ...)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}

\item{...}{additional vectors}
}
\description{
Intersect multiple vectors
}
