% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_name.R
\name{is_data_name}
\alias{is_data_name}
\title{Check if data name exists}
\usage{
is_data_name(name, dataset)
}
\arguments{
\item{name}{data name}

\item{dataset}{dataset}
}
\description{
Check if data name exists
}
