% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_origin.R
\name{is_data_origin}
\alias{is_data_origin}
\title{Checks if given data.origin is among the defaults}
\usage{
is_data_origin(data.origin)
}
\arguments{
\item{data.origin}{data origin}
}
\description{
Checks if given data.origin is among the defaults
}
