% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_observations_data.R
\name{is_observations_data}
\alias{is_observations_data}
\title{Checks if observations data exists}
\usage{
is_observations_data(name, set, dataset)
}
\arguments{
\item{name}{name of observations data}

\item{set}{observations set}

\item{dataset}{dataset}
}
\description{
Checks if observations data exists
}
