% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_observations_set.R
\name{is_observations_set}
\alias{is_observations_set}
\title{Checks if observations set exists}
\usage{
is_observations_set(set, dataset)
}
\arguments{
\item{set}{observations set}

\item{dataset}{dataset}
}
\description{
Checks if observations set exists
}
