% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_variables_data.R
\name{is_variables_data}
\alias{is_variables_data}
\title{Checks if variables data name exist}
\usage{
is_variables_data(name, dataset)
}
\arguments{
\item{name}{name of variables data}

\item{dataset}{dataset}
}
\description{
Checks if variables data name exist
}
