% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2group_vector.R
\name{list2group_vector}
\alias{list2group_vector}
\title{Transforms list to group vector for functional enrichment}
\usage{
list2group_vector(x)
}
\arguments{
\item{x}{list of proein ids}
}
\description{
Transforms list to group vector for functional enrichment
}
