% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_handling_helpers.R
\name{list2tibble}
\alias{list2tibble}
\title{Transforms a list to tibble logical indication for variables}
\usage{
list2tibble(x, identifier = "variables")
}
\arguments{
\item{x}{list}
}
\description{
Transforms a list to tibble logical indication for variables
}
