% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_entries2vector.R
\name{list_entries2vector}
\alias{list_entries2vector}
\title{Transforms list entries to vector of one element each; specific for rawdata2list}
\usage{
list_entries2vector(x)
}
\arguments{
\item{x}{data.list}
}
\description{
Transforms list entries to vector of one element each; specific for rawdata2list
}
