% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{Returns all files in the Data folder excluding RData}
\usage{
list_files(dir, exclude = c("README", ".RData"))
}
\arguments{
\item{dir}{directory of files}

\item{exclude}{expression based on which to remove files}
}
\description{
Returns all files in the Data folder excluding RData
}
