% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_msigdb_data.R
\name{load_msigdb_data}
\alias{load_msigdb_data}
\title{Extracts category specific TERM2GENE data frames from local MSigDB data}
\usage{
load_msigdb_data(category, subcategory, dataset)
}
\arguments{
\item{category}{category (see MSig_categories)}

\item{dataset}{dataset}
}
\description{
Extracts category specific TERM2GENE data frames from local MSigDB data
}
