% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variables.R
\name{map_variables}
\alias{map_variables}
\title{Return variables by matching variables.data}
\usage{
map_variables(variables.data, match.to, dataset)
}
\arguments{
\item{variables.data}{named vector or list of variables data to match}

\item{match.to}{column provided variables.data should be matched to}

\item{dataset}{dataset}
}
\description{
Return variables by matching variables.data
}
