% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_v2l.R
\name{match_v2l}
\alias{match_v2l}
\title{Matches vector entries to list (stackoverflow question 11002391)}
\usage{
match_v2l(vector, list)
}
\arguments{
\item{vector}{vector from which to match}

\item{list}{list vector entries should be matched to}
}
\description{
Matches vector entries to list (stackoverflow question 11002391)
}
