% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_rows.R
\name{matrix_rows}
\alias{matrix_rows}
\title{Subsets matrix by rownames and checks for transformations when only one row is selected}
\usage{
matrix_rows(matrix, row.names)
}
\arguments{
\item{matrix}{matrix}

\item{row.names}{row naems}
}
\description{
Subsets matrix by rownames and checks for transformations when only one row is selected
}
