% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_TERM2GENE.R
\name{merge_TERM2GENE}
\alias{merge_TERM2GENE}
\title{Combines TERM2GENE dataframes}
\usage{
merge_TERM2GENE(...)
}
\arguments{
\item{...}{TERM2GENE dataframes}
}
\description{
Combines TERM2GENE dataframes
}
