% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common_character.R
\name{most_common_character}
\alias{most_common_character}
\title{Finds and returns most common character in a data}
\usage{
most_common_character(x, exclude = NA)
}
\arguments{
\item{x}{data frame}

\item{exclude}{characters to be excluded from list}
}
\description{
Finds and returns most common character in a data
}
