% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{new_analysis_list}
\alias{new_analysis_list}
\title{Add or replace analysis list}
\usage{
new_analysis_list(silent = F, replace = F)
}
\arguments{
\item{silent}{disable messages}

\item{replace}{replace if list already exists}
}
\description{
Add or replace analysis list
}
