% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{new_cache}
\alias{new_cache}
\title{Creates new cache list}
\usage{
new_cache(replace = T)
}
\arguments{
\item{replace}{Should cache list be replaced if already existing}
}
\value{
whether a new cache list was created
}
\description{
Creates new cache list
}
