% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_create_datasets.R
\name{new_dataset}
\alias{new_dataset}
\title{Sorts raw data into data list with attributes}
\usage{
new_dataset(
  import,
  name,
  data.type,
  identifier,
  modify.identifiers = "not",
  data.columns
)
}
\arguments{
\item{import}{imported raw data frame}

\item{name}{name of dataset}

\item{data.type}{output type from MaxQuant}

\item{identifier}{(optional) specific vector of identifier column/s}

\item{modify.identifiers}{action to perform on identifier column ("not" for
nothing; "split" to separate strings by separator)}

\item{data.columns}{(optional) types of data to extract; uses defaults if
not specified}
}
\description{
Sorts raw data into data list with attributes
}
