% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{new_datasets_list}
\alias{new_datasets_list}
\title{Creates datasets list}
\usage{
new_datasets_list(replace = F)
}
\arguments{
\item{replace}{Should datasets list be replaced if already existing}
}
\description{
Creates datasets list
}
