% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{new_imports_list}
\alias{new_imports_list}
\title{Creates imports list}
\usage{
new_imports_list(replace = F)
}
\arguments{
\item{replace}{Should imports list be replaced if already existing}
}
\description{
Creates imports list
}
