% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_internal_data_structure.R
\name{new_info_list}
\alias{new_info_list}
\title{Creates new info list}
\usage{
new_info_list(replace = F)
}
\arguments{
\item{replace}{Should info list be replaced if already existing}
}
\value{
logical indication if new .info list was created
}
\description{
Creates new info list
}
