% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalizes and returns data}
\usage{
normalize(data, method = "pqn")
}
\arguments{
\item{data}{data}

\item{method}{method to use (default = "pqn", "none")}
}
\description{
Normalizes and returns data
}
