% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2g.R
\name{p2g}
\alias{p2g}
\title{Returns gene symbol from known proteins in dataset}
\usage{
p2g(proteins, OrgDb, taxId, dataset)
}
\arguments{
\item{proteins}{UniProt protein Ids}

\item{OrgDb}{string of Annotation package name to use}

\item{taxId}{taxonomy ID}

\item{dataset}{dataset}
}
\description{
Returns gene symbol from known proteins in dataset
}
