% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_terms_overlap_heatmap.R
\name{paste_terms_overlap_heatmap}
\alias{paste_terms_overlap_heatmap}
\title{Combines names and count of annotation terms}
\usage{
paste_terms_overlap_heatmap(terms, max = 5, include.count = F)
}
\arguments{
\item{terms}{named term vector}

\item{max}{maximum number of terms}

\item{include.count}{incude count in brackets}
}
\description{
Combines names and count of annotation terms
}
