% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots PCA with ggplot2}
\usage{
plot_pca(
  data_,
  x = "PC1",
  y = "PC2",
  include.variance = T,
  include.ellipses = F,
  color.by,
  color,
  fill.by,
  fill,
  shape.by,
  shape,
  size.by,
  size,
  point.transparency = 1,
  ellipse.size = 1,
  ellipse.transparency = 1,
  ellipse.linetype = 1,
  custom.theme = theme_thesis_framed_no_axis,
  aspect.ratio = 1,
  plot.center,
  axis.unit.ratio,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0),
  x.axis.breaks = 1,
  y.axis.breaks = 1,
  legend.title.color,
  legend.title.shape,
  legend.title.fill,
  legend.title.size,
  legend.position = "right",
  legend.rows,
  view = T,
  input = "data_pca",
  output = "plot_pca"
)
}
\arguments{
\item{data_}{data_}

\item{x}{Principal component to plot on x-axis}

\item{y}{Principal component to plot on y-axis}

\item{include.variance}{add variance to axis titles}

\item{include.ellipses}{add ellipses around data points}

\item{color.by}{column to use for coloring data points}

\item{color}{color}

\item{fill.by}{column to use for filling data points}

\item{fill}{fill (same as color if not defined)}

\item{shape.by}{column to use for shape of data points}

\item{shape}{shape}

\item{size.by}{column to use for size of data points}

\item{size}{point size (default = 1)}

\item{point.transparency}{transparency (0-1)}

\item{ellipse.size}{thiccness of ellipse lines (default = 1)}

\item{ellipse.transparency}{transparency of ellipse lines (default = 1)}

\item{ellipse.linetype}{linetype of ellipses (default = 1)}

\item{custom.theme}{theme to use for plot}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand x.axis (see scale_y_continuous)}

\item{x.axis.breaks}{space between x-axis breaks}

\item{y.axis.breaks}{space between y-axis breaks}

\item{legend.title.color}{title of color legend}

\item{legend.title.shape}{title of shape legend}

\item{legend.title.fill}{title of fill legend}

\item{legend.title.size}{title of size legend}

\item{legend.position}{position of legend}

\item{legend.rows}{(optional) number of rows legend content should be
presented in}

\item{view}{view plot}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Plots PCA with ggplot2
}
