% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn.R
\name{pqn}
\alias{pqn}
\title{Probabilistic quotient normalization}
\usage{
pqn(data, data2)
}
\arguments{
\item{data}{data to be normalized}

\item{data2}{(optional) data from which normalization factors can be calculated}
}
\description{
Probabilistic quotient normalization
}
