% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepares data for analysis}
\usage{
prep_data(
  data,
  min.0 = 0.5,
  normalize.method = "pqn",
  shift = 1.8,
  width = 0.3,
  seed = 123
)
}
\arguments{
\item{data}{data}

\item{min.0}{threshold to keep proteins}

\item{normalize.method}{method for normalization}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed}{seed for imputation}
}
\description{
Prepares data for analysis
}
