% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_options.R
\name{print_options}
\alias{print_options}
\title{Prints options similar to menu() function}
\usage{
print_options(options)
}
\arguments{
\item{options}{strings}
}
\description{
Prints options similar to menu() function
}
