% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein2function.R
\name{protein2function}
\alias{protein2function}
\title{Returns list vector of protein functions from UniProt}
\usage{
protein2function(proteins, dataset)
}
\arguments{
\item{proteins}{protein vector}

\item{dataset}{}
}
\description{
Returns list vector of protein functions from UniProt
}
