% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein2gene.R
\name{protein2gene}
\alias{protein2gene}
\title{Translates protein accession Ids to gene names}
\usage{
protein2gene(proteins, dataset)
}
\arguments{
\item{proteins}{proteins}

\item{dataset}{which dataset to use}
}
\description{
Translates protein accession Ids to gene names
}
