% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull names vector from}
\usage{
pull_data(data_, values, names, dataset, input, output, output.type)
}
\arguments{
\item{data_}{list or tibble}

\item{values}{column for vector values; by default second column}

\item{names}{column for vector names; by default first column}

\item{dataset}{dataset}

\item{input}{name of input data}

\item{output}{name of output data}

\item{output.type}{(optional) type of output ("vector" to force vector output)}
}
\description{
Pull names vector from
}
