% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{put_data}
\alias{put_data}
\title{Assemble data from dataset and return in list}
\usage{
put_data(
  data_,
  which,
  variables = "default",
  observations = "default",
  observations.set,
  dataset,
  output
)
}
\arguments{
\item{data_}{data_ list}

\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{observations.set}{set of observations}

\item{dataset}{dataset name or number}

\item{output}{name to save in list}
}
\description{
Assemble data from dataset and return in list
}
