% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_observations_data.R
\name{remove_observations_data}
\alias{remove_observations_data}
\title{Removes observations data}
\usage{
remove_observations_data(
  name,
  observations.set,
  dataset,
  require.confirmation = T
)
}
\arguments{
\item{name}{names of columns to be removed}

\item{observations.set}{observations.set}

\item{dataset}{dataset}

\item{require.confirmation}{Require additional confirmation before execution}
}
\description{
Removes observations data
}
