% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_processing.R
\name{remove_variables}
\alias{remove_variables}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
remove_variables(data_, min = 0.5, input, output)
}
\arguments{
\item{data_}{data}

\item{min}{min. fraction of values above 0}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\description{
Removes columns from data frame based on minimum values above 0
}
