% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_variables_data.R
\name{remove_variables_data}
\alias{remove_variables_data}
\title{Removes variables data}
\usage{
remove_variables_data(name, dataset, require.confirmation = T)
}
\arguments{
\item{name}{names of columns to be removed}

\item{dataset}{dataset}

\item{require.confirmation}{Require additional confirmation before execution}
}
\description{
Removes variables data
}
