% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_data.R
\name{reorder_data}
\alias{reorder_data}
\title{Reorders dataframe rows according to observations data order}
\usage{
reorder_data(data.name, observations.set, dataset, silent = F)
}
\arguments{
\item{data.name}{name of data}

\item{observations.set}{observations set}

\item{dataset}{dataset}

\item{silent}{return error if function fails}
}
\description{
Reorders dataframe rows according to observations data order
}
