% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_prefix.R
\name{same_prefix}
\alias{same_prefix}
\title{Test vector of strings for a common prefix of minimal given length}
\usage{
same_prefix(names, min.similarity = 8, enforce.residue = T)
}
\arguments{
\item{names}{vector of strings}

\item{min.similarity}{minimal similarity of prefix}

\item{enforce.residue}{is a character after the prefix required}
}
\value{
TRUE if common prefix exists
}
\description{
Test vector of strings for a common prefix of minimal given length
}
