% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2cache.R
\name{save2cache}
\alias{save2cache}
\title{Saves data in cache}
\usage{
save2cache(data, view = F, new = F)
}
\arguments{
\item{data}{date}

\item{view}{Should data be viewed after saving}

\item{new}{Should cache list be renewed before saving}
}
\description{
Saves data in cache
}
