% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Store analysis object in Analysis list}
\usage{
save_results(data, name, replace)
}
\arguments{
\item{data}{data to be saved}

\item{name}{name of data}

\item{replace}{replace existing element with same name}
}
\description{
Store analysis object in Analysis list
}
