% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wd.R
\name{save_wd}
\alias{save_wd}
\title{Saves given or current working directory}
\usage{
save_wd(wd, silent = F)
}
\arguments{
\item{wd}{working directory}

\item{silent}{Should a message be}
}
\description{
Saves given or current working directory
}
