% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_vector2list.R
\name{sep_vector2list}
\alias{sep_vector2list}
\title{# Generates a list from separated vector}
\usage{
sep_vector2list(x, sep)
}
\arguments{
\item{x}{vector}

\item{sep}{separator}
}
\description{
# Generates a list from separated vector
}
