% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_continuous_axes.R
\name{set_continuous_axes}
\alias{set_continuous_axes}
\title{Defines ggplot panel by ratio, size, unit size, center, and axis breaks}
\usage{
set_continuous_axes(
  p,
  x.axis.limits,
  y.axis.limits,
  aspect.ratio = 1,
  plot.center,
  axis.unit.ratio,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0),
  x.axis.breaks = 1,
  y.axis.breaks = 1
)
}
\arguments{
\item{p}{plot}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand x.axis (see scale_y_continuous)}

\item{x.axis.break.size}{distance between x-axis breaks}

\item{y.axis.break.size}{distance between y-axis breaks}
}
\description{
Defines ggplot panel by ratio, size, unit size, center, and axis breaks
}
