% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dataset_attr.R
\name{set_dataset_attr}
\alias{set_dataset_attr}
\title{Sets dataset attribute; does not check if exists}
\usage{
set_dataset_attr(x, which, dataset)
}
\arguments{
\item{x}{attribute}

\item{which}{attribute type}

\item{dataset}{dataset}
}
\description{
Sets dataset attribute; does not check if exists
}
