% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_data_name.R
\name{set_default_data_name}
\alias{set_default_data_name}
\title{Set default data name}
\usage{
set_default_data_name(name, dataset)
}
\arguments{
\item{name}{data name}

\item{dataset}{dataset}
}
\description{
Set default data name
}
