% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_dataset.R
\name{set_default_dataset}
\alias{set_default_dataset}
\title{Updates default dataset}
\usage{
set_default_dataset(name, silent = F)
}
\arguments{
\item{name}{Name of dataset}

\item{silent}{suppress messages}
}
\description{
Updates default dataset
}
