% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_observations_set.R
\name{set_default_observations_set}
\alias{set_default_observations_set}
\title{Set default data set}
\usage{
set_default_observations_set(set, dataset, silent = F)
}
\arguments{
\item{set}{observations set name}

\item{dataset}{dataset}

\item{silent}{Silent?}
}
\description{
Set default data set
}
