% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diagonal.R
\name{set_diagonal}
\alias{set_diagonal}
\title{Sets diagonal values of symmetric matrix}
\usage{
set_diagonal(data, diag = 1)
}
\arguments{
\item{data}{data}

\item{diag}{new value for diagonal values}
}
\description{
Sets diagonal values of symmetric matrix
}
