% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_project_setup.R
\name{setup_folders}
\alias{setup_folders}
\title{Creates folder structure for analysis project}
\usage{
setup_folders(dir, silent = F)
}
\arguments{
\item{dir}{directory of project}

\item{silent}{suppress message when folders created}
}
\description{
Creates folder structure for analysis project
}
