% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant.R
\name{significant}
\alias{significant}
\title{Returns logical vector indication significant p-values}
\usage{
significant(
  p.values,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  na.value = NA
)
}
\arguments{
\item{p.values}{named vector of p-values}

\item{pvalueCutoff}{limit for p.values and FDR for Benjamini-Hochberg-Correction}

\item{pAdjustMethod}{method to determine significant p-values ("none", "BH" (Benjamini-Hochberg-Correction)); see p.adjust.methods}

\item{na.value}{return value for NA}
}
\description{
Returns logical vector indication significant p-values
}
