% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculates similarity from correlation}
\usage{
similarity(data, method = "none")
}
\arguments{
\item{data}{correlation matrix}

\item{method}{similarity function ("absolute", "preserve", "none")}
}
\description{
Calculates similarity from correlation
}
