% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_enrichment_results.R
\name{simplify_enrichment_results}
\alias{simplify_enrichment_results}
\title{Simplifies enrichment data frame by overlap between annotations, semantic similarity for GO terms and minimum protein count}
\usage{
simplify_enrichment_results(
  enrichment_df,
  by.overlap = T,
  overlap.threshold = 0.7,
  separator = ";",
  by.semantic = F,
  similarity.measure = "Wang",
  similarity.threshold = 0.4,
  dataset,
  overlap.first = T,
  minimum.count = 0
)
}
\arguments{
\item{enrichment_df}{}

\item{by.overlap}{use simplify_set_overlap}

\item{overlap.threshold}{minimum fraction of common proteins in protein set with lower p-value to be excluded}

\item{separator}{string separator of proteins}

\item{by.semantic}{use simplify_GO_terms_semantic}

\item{similarity.measure}{method to measure pairwise GO terms similarity (see ?GOSemSim::mgoSim)}

\item{similarity.threshold}{threshold to remove terms (0-1, default = 0.4)}

\item{dataset}{dataset}

\item{minimum.count}{number of proteins per annotation}
}
\description{
Simplifies enrichment data frame by overlap between annotations, semantic similarity for GO terms and minimum protein count
}
