% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_set_overlap.R
\name{simplify_set_overlap}
\alias{simplify_set_overlap}
\title{Removes similar protein sets based on minimum overlap}
\usage{
simplify_set_overlap(proteins, overlap.threshold = 0.7, separator = ";")
}
\arguments{
\item{proteins}{vector containing protein strings separated by ;}

\item{overlap.threshold}{minimum fraction of common proteins in protein set with lower p-value to be excluded}

\item{separator}{string separator of proteins}
}
\value{
Names logical vector
}
\description{
Removes similar protein sets based on minimum overlap
}
