% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_functions.R
\name{strsplit_}
\alias{strsplit_}
\title{Same as strsplit but option to return as vector and to convert to numeric}
\usage{
strsplit_(x, split, output.type = "vector", as_numeric = T)
}
\arguments{
\item{x}{string}

\item{split}{split character}

\item{output.type}{output type (default = "vector"; other option is "list")}

\item{as.numeric}{try to convert to numeric}
}
\description{
Same as strsplit but option to return as vector and to convert to numeric
}
