% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_functions.R
\name{strsplit_keep}
\alias{strsplit_keep}
\title{Performs strsplit and keeps elements specified by vector}
\usage{
strsplit_keep(x, keep = 1, split, as_numeric = T)
}
\arguments{
\item{x}{vector}

\item{keep}{vector of elements to keep}

\item{split}{separator}

\item{as_numeric}{try to convert to numeric}
}
\description{
Performs strsplit and keeps elements specified by vector
}
