% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_observations.R
\name{subset_observations}
\alias{subset_observations}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
subset_observations(
  data_,
  observations,
  observations.set,
  dataset,
  input,
  output
)
}
\arguments{
\item{data_}{data}

\item{observations}{observarions}

\item{observations.set}{observations.set}

\item{dataset}{dataset}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\description{
Removes columns from data frame based on minimum values above 0
}
