% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data_construction.R
\name{subset_variables}
\alias{subset_variables}
\title{Removes columns from data frame based on minimum values above 0}
\usage{
subset_variables(data_, variables, dataset, input, output)
}
\arguments{
\item{data_}{data}

\item{variables}{variables}

\item{dataset}{dataset}

\item{input}{if data_ is list: name of data to use}

\item{output}{if data_ is list: name of output data to save in list under}
}
\description{
Removes columns from data frame based on minimum values above 0
}
