% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_proteins.R
\name{summary_proteins}
\alias{summary_proteins}
\title{Views a list of proteins}
\usage{
summary_proteins(proteins, sort.names = F, view = T)
}
\arguments{
\item{proteins}{vector of proteins}

\item{sort.names}{order table by protein names}

\item{view}{view protein summary table}
}
\description{
Views a list of proteins
}
