% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_proteins_2.R
\name{summary_proteins_2}
\alias{summary_proteins_2}
\title{Views a list of proteins with links to UniProt}
\usage{
summary_proteins_2(proteins, caption = NULL, sort.names = F)
}
\arguments{
\item{proteins}{vector of proteins}

\item{caption}{table caption}

\item{sort.names}{order table by protein names}
}
\description{
Views a list of proteins with links to UniProt
}
