% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2g.R
\name{t2g_get_genes}
\alias{t2g_get_genes}
\title{Get vector or list of genes mapped to given term/s}
\usage{
t2g_get_genes(T2G, terms, output.type)
}
\arguments{
\item{T2G}{term-to-gene data frame}

\item{terms}{character vector of terms}

\item{output.type}{"vector" (default for one term) or "list" (default for
multiple terms)}
}
\description{
Get vector or list of genes mapped to given term/s
}
