% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2g.R
\name{t2g_get_matrix}
\alias{t2g_get_matrix}
\title{Get tibble of terms mapped to genes}
\usage{
t2g_get_matrix(T2G, terms = NULL, genes = NULL, output.type = "tibble")
}
\arguments{
\item{T2G}{term-to-gene data frame}

\item{terms}{character vector of terms}

\item{genes}{character vector of genes}

\item{output.type}{"tibble" (default) or "matrix"}
}
\description{
Get tibble of terms mapped to genes
}
