% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2g.R
\name{t2g_get_terms}
\alias{t2g_get_terms}
\title{Get vector or list of terms mapped to given gene/s}
\usage{
t2g_get_terms(T2G, genes, output.type)
}
\arguments{
\item{T2G}{term-to-gene data frame}

\item{genes}{character vector of genes}

\item{output.type}{"vector" (default for one gene) or "list" (default for
multiple genes)}
}
\description{
Get vector or list of terms mapped to given gene/s
}
