% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2g.R
\name{t2g_subset_genes}
\alias{t2g_subset_genes}
\title{Subset term-to-gene data frame by given genes}
\usage{
t2g_subset_genes(T2G, genes)
}
\arguments{
\item{T2G}{term-to-gene data frame}

\item{genes}{character vector contaiing genes to keep}
}
\description{
Subset term-to-gene data frame by given genes
}
