% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2g.R
\name{t2g_subset_terms}
\alias{t2g_subset_terms}
\title{Subset term-to-gene data frame by given terms}
\usage{
t2g_subset_terms(T2G, terms)
}
\arguments{
\item{T2G}{term-to-gene data frame}

\item{terms}{character vector containing terms to keep}
}
\description{
Subset term-to-gene data frame by given terms
}
