% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hjv_heatmap_only}
\alias{theme_hjv_heatmap_only}
\title{Create a theme for half open plots for Hjv project}
\usage{
theme_hjv_heatmap_only(
  font_size = 8,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)
}
\arguments{
\item{font_size}{Overall font size.}

\item{font_family}{Font family for plot title, axis titles and labels, legend texts, etc.}

\item{line_size}{Line size for axis lines.}

\item{rel_small}{Relative size of small text (e.g., axis tick labels)}

\item{rel_tiny}{Relative size of tiny text (e.g., caption)}

\item{rel_large}{Relative size of large text (e.g., title)}
}
\value{
The theme.
}
\description{
Create a theme for half open plots for Hjv project
}
