% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_phosprot_half_open}
\alias{theme_phosprot_half_open}
\title{Create a theme for half open plots for Hjv project}
\usage{
theme_phosprot_half_open(
  font_size = 8,
  font_size_small = 7,
  font_size_tiny = 6,
  font_size_large = 10,
  font_family = "",
  line_size = 0.5
)
}
\arguments{
\item{font_size}{Overall font size.}

\item{font_size_small}{Relative size of small text (e.g., axis tick labels)}

\item{font_size_tiny}{Relative size of tiny text (e.g., caption)}

\item{font_size_large}{Relative size of large text (e.g., title)}

\item{font_family}{Font family for plot title, axis titles and labels,
legend texts, etc.}

\item{line_size}{Line size for axis lines.}
}
\value{
The theme.
}
\description{
Create a theme for half open plots for Hjv project
}
