% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Returns logical vector indicating values above or below given limit}
\usage{
threshold(values, min = T, limit = 0, na.value = NA)
}
\arguments{
\item{values}{numeric vector}

\item{min}{limit is a minimum/lower limit}

\item{limit}{cutoff value}

\item{na.value}{value value for NA}
}
\description{
Returns logical vector indicating values above or below given limit
}
