% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data_.R
\name{to_data_}
\alias{to_data_}
\title{Adds data to data_ list}
\usage{
to_data_(data, name = "data")
}
\arguments{
\item{data}{data}

\item{name}{name of list element}
}
\description{
Adds data to data_ list
}
