% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{transfer_data_frames}
\alias{transfer_data_frames}
\title{Transfers data from raw_dataset to .datasets}
\usage{
transfer_data_frames(dataset, data.columns)
}
\arguments{
\item{dataset}{dataset}

\item{data.columns}{data frames to transfer}
}
\description{
Transfers data from raw_dataset to .datasets
}
