% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{transfer_observations}
\alias{transfer_observations}
\title{Transfers observations from raw_dataset to .datasets}
\usage{
transfer_observations(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\description{
Transfers observations from raw_dataset to .datasets
}
