% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{transfer_variables_data}
\alias{transfer_variables_data}
\title{Transfers variables data from raw_dataset to .datasets}
\usage{
transfer_variables_data(dataset, data.columns)
}
\arguments{
\item{dataset}{dataset}

\item{data.columns}{columns to transfer}
}
\description{
Transfers variables data from raw_dataset to .datasets
}
