% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_id.R
\name{translate_Ids}
\alias{translate_Ids}
\title{Translates biological IDs using annotations packages}
\usage{
translate_Ids(
  Ids,
  fromType = "UNIPROT",
  toType,
  OrgDb,
  dataset,
  drop = F,
  fill.missing = "initial",
  save = T,
  silent = T
)
}
\arguments{
\item{Ids}{Ids}

\item{fromType}{type of given Ids}

\item{toType}{type Ids should be translated to}

\item{OrgDb}{name of organism database to use; defined by setup_annotations()}

\item{dataset}{dataset}

\item{drop}{drop missing translations}

\item{fill.missing}{How to handle missing Ids. By default, given Ids are filled in, other characters or NA can be used}

\item{save}{translated Ids in variables data frame for next time (all Ids in dataset will be translated)}

\item{silent}{suppress all messages}
}
\description{
Translates biological IDs using annotations packages
}
