% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables2labels.R
\name{variables2labels}
\alias{variables2labels}
\title{Wrapper around get_variables_data to replace variables with labels}
\usage{
variables2labels(variables, name, dataset)
}
\arguments{
\item{variables}{variables}

\item{name}{variables data name}

\item{dataset}{dataset}
}
\description{
Wrapper around get_variables_data to replace variables with labels
}
