% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data.R
\name{view_data}
\alias{view_data}
\title{Shows all data lists}
\usage{
view_data(which = c(".imports", ".info", ".databases", ".datasets"))
}
\arguments{
\item{which}{data ftrames to show}
}
\description{
Shows all data lists
}
