% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data.R
\name{view_observations}
\alias{view_observations}
\title{Prints observations data frame}
\usage{
view_observations(dataset, observations.set)
}
\arguments{
\item{dataset}{dataset}

\item{observations.set}{observations set}
}
\description{
Prints observations data frame
}
